api/requirements.txt
flask
flask-cors
python-dotenv
requests
openai
PyPDF2
python-docx
